DM-6287 chart refactoring: moved chart-related code to charts package, converted React components to ES6 classes, fixed bug when undefined shows as a label.
