+++ src/firefly/js/charts/HistogramCntlr.js
