+++ src/firefly/js/charts/TableStatsCntlr.js
